fix FCP issues finally
issues - FCP will read from CreationDate if it exists, but that field is
not TZ aware.

CreateDate and ModifyDate are TZ aware, but when Google Photos
downloads, it treats those as UTC and doesn't save the TZ information.

This is likely a Google Photos bug where it's confusing those fields
with the Quicktime fields MediaCreateDate and MediaModifyDate, which
are integers and supposed to be in UTC.

So this now fixes that by removing CreationDate, writing back TZ info
to CreateDate and ModifyDate, and writing UTC to
MediaCreateDate/MediaModifyDate.